+++ dataservices/management/commands/import_world_economic_outlook_data.py
@@ -9,27 +9,31 @@ from dataservices.models import Country, WorldEconomicOutlookByCountry
 class Command(BaseCommand):
     help = 'Import IMF world economic outlook data by country from Data Workspace'
 
-    imf_economic_growth_descriptor = 'NGDP_RPCH'
-    imf_gdp_per_capita_descriptor = 'NGDPDPC'
-
     engine = sa.create_engine(settings.DATA_WORKSPACE_DATASETS_URL, execution_options={'stream_results': True})
     sql = '''
         SELECT
-            iso AS iso3,
-            weo_subject_code AS subject_code,
+            iso AS ons_iso_alpha_3_code,
+            CASE WHEN weo_subject_code = 'NGDPD' THEN
+                'MKT_POS'
+            ELSE
+                weo_subject_code
+            END AS subject_code,
             subject_descriptor,
             subject_notes,
             units,
             scale,
             LTRIM(x.year, 'year_') AS year,
-            x.value,
+            CASE WHEN weo_subject_code = 'NGDPD' THEN
+                Rank() OVER (PARTITION BY year, weo_subject_code ORDER BY x.value::numeric DESC)
+            ELSE
+                x.value::numeric
+            END AS value,
             estimates_start_after
         FROM
             imf.world_economic_outlook__by_country AS dataset,
-            jsonb_each_text(to_jsonb (dataset)) AS x (year,
-                value)
+            jsonb_each_text(to_jsonb (dataset)) AS x (year, value)
         WHERE
-            weo_subject_code IN('NGDPDPC', 'NGDP_RPCH')
+            weo_subject_code IN('NGDPD', 'NGDPDPC', 'NGDP_RPCH')
             AND x.year LIKE 'year_%'
             AND NULLIF(TRIM(x.value), '') IS NOT NULL;
     '''
@@ -41,13 +45,14 @@ class Command(BaseCommand):
         for chunk in chunks:
             for _idx, row in chunk.iterrows():
                 try:
-                    country = Country.objects.get(iso3=row.iso3)
+                    country = Country.objects.get(iso3=row.ons_iso_alpha_3_code)
                 except Country.DoesNotExist:
-                    continue
+                    country = None
 
                 data.append(
                     WorldEconomicOutlookByCountry(
                         country=country,
+                        ons_iso_alpha_3_code=row.ons_iso_alpha_3_code,
                         subject_code=row.subject_code,
                         subject_descriptor=row.subject_descriptor,
                         subject_notes=row.subject_notes,