+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre release
 ### Enhancements
 ### Bugs fixed
+- GLS-428 - Handle missing markets from the IMF dataset
 
 ## [2.5.0](https://github.com/uktrade/directory-api/releases/tag/2.5.0)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/2.4.2...2.5.0)